issue #360: reusable LabelDataValidator component
