+++ src/components/GuaranteedAnalysisForm.tsx
@@ -10,7 +10,7 @@ const GuaranteedAnalysisForm: React.FC<FormComponentProps> = ({
   labelData,
   setLabelData,
 }) => {
-  const { t } = useTranslation("labelDataValidationPage");
+  const { t } = useTranslation("labelDataValidator");
   const methods = useForm<LabelData>({
     defaultValues: labelData,
   });
